pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "superset-hetuengine-connector"
version = "0.1.0"
description = "Apache Superset database connector for Huawei HetuEngine"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Apache-2.0" }
authors = [
    { name = "Community Contributors" },
]
keywords = [
    "superset",
    "hetuengine",
    "huawei",
    "trino",
    "presto",
    "jdbc",
    "database",
    "connector",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "apache-superset>=2.0.0",
    "jaydebeapi>=1.2.3",
    "JPype1>=1.4.0",
    "sqlalchemy>=1.4.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
]

[project.urls]
Homepage = "https://github.com/pesnik/superset-hetuengine-connector"
Documentation = "https://github.com/pesnik/superset-hetuengine-connector/blob/main/README.md"
Repository = "https://github.com/pesnik/superset-hetuengine-connector"
"Bug Reports" = "https://github.com/pesnik/superset-hetuengine-connector/issues"
"Source Code" = "https://github.com/pesnik/superset-hetuengine-connector"

[project.entry-points."sqlalchemy.dialects"]
hetuengine = "superset_hetuengine.sqlalchemy_dialect:HetuEngineDialect"

[project.entry-points."superset.db_engine_specs"]
hetuengine = "superset_hetuengine.db_engine_spec:HetuEngineSpec"

[tool.setuptools]
packages = ["superset_hetuengine"]

[tool.setuptools.package-data]
superset_hetuengine = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=superset_hetuengine",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["superset_hetuengine"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 88
target-version = "py38"
extend-exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A"]
ignore = ["E203", "E501"]
